Recreate in Typescript
