+++ app.html
@@ -10,6 +10,6 @@
     
     <div id="root"></div>
     
-    <script type="module" src="/src/app.jsx"></script>
+    <script type="module" src="/src/app.tsx"></script>
   </body>
 </html>
