+++ index.html
@@ -34,6 +34,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/src/index.js"></script>
+    <script type="module" src="/src/index.ts"></script>
   </body>
 </html>
