--- jsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "typeRoots": ["./node_modules/@types", "./node_modules/@mirohq"]
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
