+++ package.json
@@ -14,6 +14,11 @@
   },
   "devDependencies": {
     "vite": "3.0.3",
-    "@vitejs/plugin-react": "^2.2.0"
+    "@mirohq/websdk-types": "latest",
+    "@types/react": "^18.0.24",
+    "@types/react-dom": "^18.0.8",
+    "@vitejs/plugin-react": "^2.2.0",
+    "typescript": "4.9.5",
+    "@types/node": "^18.8.2"
   }
 }
