+++ src/index.ts
@@ -1,0 +1,7 @@
+export async function init() {
+  miro.board.ui.on('icon:click', async () => {
+    await miro.board.ui.openPanel({url: 'app.html'});
+  });
+}
+
+init();
